Added ReducedSpherical representations and transforms. ReducedSpherical is used when you want spherical coordinates, but the problem is symmetric in phi (and thus m).
Because there is no phi or m dependence, we can use a proper gauss-legendre quadrature
for transforming between spherical harmonics and grid. It is possible to do this for 
general m, but currently it is only implemented for m=0.

A note on potential evaluation:
It is also worth noting that because we do not have phi or m in our representation, the trick 
with using higher rank on the potential than on the wavefunction (as used in the normal spherical
representation) is not needed. One must therefore not use SphericalDynamicPotentialEvaluator, but
rather the normal DynamicPotentialEvaluator when compiling potentials for ReducedSpherical problems.




git-svn-id: http://pyprop.googlecode.com/svn/trunk@66 354cffb3-6528-0410-8d42-45ce437ad3c5